+++ src/components/App/App.js
@@ -5,9 +5,7 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import Layout from '../Layout';
 import { createMuiTheme } from '@material-ui/core/styles';
 import { ThemeProvider } from '@material-ui/styles';
-import green from '@material-ui/core/colors/green';
 import lightBlue from '@material-ui/core/colors/lightBlue';
-import Location from '../Location/Location'
 import Setup from '../Setup/Setup';
 import stepperData from '../Setup/setup-stepper-data.json';
 import {getPDdata} from '../../utils';