+++ src/components/Timer/Timer.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import Timer from './Timer';
 
-describe('<App />', () => {
+describe('<Timer />', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<App />, div);
+    ReactDOM.render(<Timer />, div);
     ReactDOM.unmountComponentAtNode(div);
   });
 })