+++ src/utils/constants.js
@@ -1,1 +1,2 @@
 export const P_EMAIL = "padachone_email";
+export const BING_API = "ArNqsYDx-rtxpMbR4ddz8SyY4-dv8-JK35KErFW3GIU7_UwgaCVz8Bj4iKy4X-JP";